// src/shared/dataprovider/api/model/cardTransactionDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CashInOutStatusEnum } from './cashInOutStatusEnum';
import { TransactionTypeEnum } from './transactionTypeEnum';
import { WithdrawalStatusEnum } from './withdrawalStatusEnum';
import { WithdrawalTypeEnum } from './withdrawalTypeEnum';

export interface CardTransactionDTO { 
  
    Id?: number;
    TransactionNumber?: string;
    ReferenceNumber?: string;
    TransactionType?: TransactionTypeEnum;
    TransactionTypeDescription?: string;
    Method?: WithdrawalTypeEnum;
    MethodDescription?: string;
    Amount?: number;
    TransactionFee?: number;
    TransactionFeeOnTop?: number;
    TransactionFeeAtCost?: number;
    NetAmount?: number;
    Status?: WithdrawalStatusEnum;
    StatusDescription?: string;
    CreatedDate?: Date;
    CompletedDate?: Date;
    RequesterUserId?: number;
    RequesterName?: string;
    CardIdDestination?: number;
    CardNameDestination?: string;
    Description?: string;
    ApproverName?: string;
}